# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Monitor de Ordens Planejadas SAP
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ordem {
  ordem: string;
  op?: string;
  material: string;
  unid: string;
  qtd: number;
  qtdEntrada: number;
  qtdPlan: number;
  desvio: number;
  valor: number;
  dtAbertura?: string;
  dtInicio?: string;
  dtRemessaPlan?: string;
  dtRemessaReal?: string;
  recebConcluido: boolean;
  dataFim?: string;
  centro: string;
  tipoOrdem: string;
  sincronizadoEm?: string;
}

export interface OrdersResponse {
  ordens: Ordem[];
  total: number;
  limit: number;
  offset: number;
}

export interface SyncOrdersPayload {
  /** @minItems 1 */
  ordens: Ordem[];
  fonte?: string;
  timestamp?: string;
}

export interface SyncOrdersResponse {
  success: boolean;
  processados: number;
  fonte?: string;
  timestamp: string;
}

export interface SapSyncStatus {
  isRunning?: boolean;
  lastSync?: string;
  nextSync?: string;
  success?: boolean;
  message?: string;
  duration?: number;
  timestamp?: string;
}

export interface ErrorResponse {
  error: string;
  success?: boolean;
}

export type ListOrdersParams = {
centro?: string;
material?: string;
tipoOrdem?: string;
busca?: string;
/**
 * @minimum 1
 * @maximum 5000
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};