# lib/api-zod/src/generated/types/ordem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Monitor de Ordens Planejadas SAP
 * OpenAPI spec version: 0.2.0
 */

export interface Ordem {
  ordem: string;
  op?: string;
  material: string;
  unid: string;
  qtd: number;
  qtdEntrada: number;
  qtdPlan: number;
  desvio: number;
  valor: number;
  dtAbertura?: string;
  dtInicio?: string;
  dtRemessaPlan?: string;
  dtRemessaReal?: string;
  recebConcluido: boolean;
  dataFim?: string;
  centro: string;
  tipoOrdem: string;
  sincronizadoEm?: string;
}