# lib/api-zod/src/generated/types/sapSyncStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Monitor de Ordens Planejadas SAP
 * OpenAPI spec version: 0.2.0
 */

export interface SapSyncStatus {
  isRunning?: boolean;
  lastSync?: string;
  nextSync?: string;
  success?: boolean;
  message?: string;
  duration?: number;
  timestamp?: string;
}