# lib/api-zod/src/generated/types/syncOrdersPayload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Monitor de Ordens Planejadas SAP
 * OpenAPI spec version: 0.2.0
 */
import type { Ordem } from './ordem';

export interface SyncOrdersPayload {
  /** @minItems 1 */
  ordens: Ordem[];
  fonte?: string;
  timestamp?: string;
}