# lib/api-zod/src/generated/types/syncOrdersResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Monitor de Ordens Planejadas SAP
 * OpenAPI spec version: 0.2.0
 */

export interface SyncOrdersResponse {
  success: boolean;
  processados: number;
  fonte?: string;
  timestamp: string;
}